Tue, Mar 21, 2023, 3:18 PM +01:00

+++ week4/filter-less/helpers.c
@@ -7,7 +7,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     for(int i = 0; i < height - 1; i++)
         for(int j = 0; j < width - 1; j++)
         {
-            int x = round((image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed) / 3.0);
+            float x = round((image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed) / 3.0);
             image[i][j].rgbtBlue = x;
             image[i][j].rgbtGreen = x;
             image[i][j].rgbtRed = x;
